requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
autopep8==2.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contexttimer==0.3.3
contourpy==1.2.0
cpu-scheduling-simulator==1.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-language==2.13.3
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.3.3
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
msgpack==1.0.8
multidict==6.0.5
mvn==0.0.1
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy @ file:///private/tmp/numpy-20240206-5894-1wazl8/numpy-1.26.4
openai==0.28.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.21.0
pooch==1.8.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
pyrate-limiter==3.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sec-api==1.0.18
sec-edgar-downloader==5.0.2
Send2Trash==1.8.2
setuptools==69.5.1
simpleaudio==1.0.4
simpy==4.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
streamlit==1.33.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
textdistance==4.6.2
tf_keras==2.16.0
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.40.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
utils==1.0.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.2_1/libexec/wheel-0.42.0-py3-none-any.whl#sha256=5f1ca525bef7466225922019d72cb06a9e91cd3ff807ddd4be1acb84655e5107
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.9.4
